modified and added some new tests
